Reorganization of Streamlit components to allow for better reusability.
